Updated and renamed verifyInteraction.js to verifyInteraction.ts
